+++ projects/ngx-forms-inputs/src/inputs/color-input/color-input.component.ts
@@ -1,6 +1,6 @@
-import {ChangeDetectionStrategy, ChangeDetectorRef, Component, Host, Input, Optional, SkipSelf} from '@angular/core';
-import {ControlContainer, FormsModule} from "@angular/forms";
-import {BaseInputComponent, FormContext} from '@consensus-labs/ngx-forms';
+import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
+import {FormsModule} from "@angular/forms";
+import {BaseInputComponent} from '@consensus-labs/ngx-forms';
 import {harmonicaAnimation, IconDirective, NoClickBubbleDirective} from '@consensus-labs/ngx-tools';
 import {ColorPickerModule} from "ngx-color-picker";
 import {AsyncPipe, NgIf} from "@angular/common";
@@ -31,13 +31,7 @@ export class ColorInputComponent extends BaseInputComponent<string | undefined, 
 
   @Input() noAlpha = false;
 
-  constructor(
-    changes: ChangeDetectorRef,
-    @Optional() @Host() @SkipSelf() controlContainer: ControlContainer,
-    @Optional() formScope: FormContext
-  ) {
-    super(changes, controlContainer, formScope);
-  }
+
 
   postprocessValue(value?: string) {
     if (!value) return undefined;