+++ projects/ngx-forms/src/lib/ngx-forms-tools/index.ts
@@ -1,5 +1,6 @@
 
 export * from './models/base-input-component';
+export * from './models/base-select-input-component';
 export * from './models/error-state-matchers';
 export * from './components/form-wrapper/form-wrapper.component';
 export * from './services/form-context.service';